Move tags declaration to their own file
